Refactore cart codes form localStorage NOTE: 1. getCart(), saveCart() & updateCartCount() functions in cart.handler.js 2. Remove some unneccessary comments in auth.js file
